pyproject.toml
[tool.poetry]
name = "better-club-management-system"
version = "0.1.0"
description = "A simple and modular, easy to configure club management system"
authors = ["Aditya Jyoti <aj.adityajyoti@gmail.com>"]
readme = "README.md"
packages = [{include = "bcms"}]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
prisma = "^0.15.0"
setuptools = "^75.0.0"
pydantic = "^2.9.1"
uuid = "^1.30"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
tox = "^4.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
format = "scripts.format_project:format_project"
update_prisma = "scripts.update_prisma:update_prisma"
